+++ frontend/src/components/customer/noti/index.ts
@@ -1,0 +1,2 @@
+export * from "./NotiBox";
+export * from "./NoNotiBox";
